+++ test/unit/src/http/http-res-fixtures.js
@@ -17,10 +17,76 @@ let xml = '<hi>there</hi>'
 
 let arc6 = {}
 
+arc6.http = {
+  // Not returning is valid, and returns a 'null' string as JSON (lol)
+  noReturn: undefined,
+
+  // ... while this sends back 0 content-length JSON
+  emptyReturn: '',
+
+  // Sending back JSON-serializable JS primitives actually coerces strings
+  string: 'hi',
+  object: { ok: true },
+  array: [ 'howdy' ],
+  buffer: Buffer.from('hi'),
+  number: 42,
+
+  // Implicit JSON returns one no longer work without statuscode + headers
+  bodyOnly: {
+    body: text
+  },
+  bodyWithStatus: {
+    statusCode: 200,
+    body: text
+  },
+
+  // Explicit return is the now hotness
+  bodyWithStatusAndContentType: {
+    statusCode: 200,
+    headers: { 'content-type': 'application/json' },
+    body: text
+  },
+
+  // Any properly base64 encoded response
+  encodedWithBinaryType: {
+    statusCode: 200,
+    body: b64enc('hi there\n'),
+    headers: { 'content-type': 'application/pdf' },
+    isBase64Encoded: true
+  },
+
+  // Set cookies param
+  cookies: {
+    statusCode: 200,
+    cookies: [ 'foo', 'bar' ],
+    body: text
+  },
+
+  // ... now SSL
+  secureCookies: {
+    statusCode: 200,
+    cookies: [ 'hi=there; Secure', 'hi=there; Secure' ],
+    body: text
+  },
+
+  // ... now via header
+  secureCookieHeader: {
+    statusCode: 200,
+    headers: { 'set-cookie': 'hi=there; Secure' },
+    body: text
+  },
+
+}
+
 arc6.rest = {
   /**
    * New params introduced with Arc 6+ APG-proxy-Lambda
    */
+  // Set body (implicit JSON return)
+  body: {
+    body: text
+  },
+
   // Set isBase64Encoded (not technically new, but implemented differently)
   isBase64Encoded: {
     body: b64enc('hi there\n'),
@@ -32,31 +98,61 @@ arc6.rest = {
     body: Buffer.from('hi there\n'),
   },
 
-  // Base64 encoded with valid binary content type
-  encodedWithBinaryType: {
+  // Improperly formed response: base64 encoded with valid binary content type
+  encodedWithBinaryTypeBad: {
     body: b64enc('hi there\n'),
     headers: { 'Content-Type': 'application/pdf' }
+  },
+
+  // Properly formed response: base64 encoded with valid binary content type
+  encodedWithBinaryTypeGood: {
+    body: b64enc('hi there\n'),
+    headers: { 'Content-Type': 'application/pdf' },
+    isBase64Encoded: true
+  },
+
+  // Set cookie via header
+  secureCookieHeader: {
+    body: html,
+    headers: { 'set-cookie': 'hi=there; Secure' }
+  },
+
+  // ... now with a multi-value header
+  secureCookieMultiValueHeader: {
+    body: html,
+    multiValueHeaders: {
+      'set-cookie': [ 'hi=there; Secure', 'hi=there; Secure' ]
+    }
   },
 
   // Set multiValueHeaders
   multiValueHeaders: {
     headers: { 'Content-Type': 'text/plain', 'Set-Cookie': 'Baz' },
-    multiValueHeaders: { 'Content-Type': [ 'text/plain' ], 'Set-Cookie': [ 'Foo', 'Bar' ] }
+    multiValueHeaders: {
+      'Content-Type': [ 'text/plain' ],
+      'Set-Cookie': [ 'Foo', 'Bar' ]
+    }
+  },
+
+  invalidMultiValueHeaders: {
+    multiValueHeaders: {
+      'Content-Type': 'text/plain',
+      'Set-Cookie': {
+        'Foo': 'Bar'
+      }
+    }
   }
 }
 
 let arc5 = {
-  /**
-   * Arc Functions response format
-   */
   // Set body
   body: {
-    body: 'hi there'
+    body: text
   },
 
   // Set cacheControl
   cacheControl: {
-    body: 'hi there',
+    body: text,
     cacheControl: 'max-age=1',
     headers: { 'cache-control': 'max-age=60' } // cacheControl should win
   },
@@ -73,7 +169,7 @@ let arc5 = {
     headers: { 'Content-Type': 'application/json; charset=uft8' }
   },
 
-  // Test default anti-caching on JSON API
+  // Test default anti-caching on JSON API – works in Functions, but never worked in V5
   noCacheControlJSONapi: {
     body: json,
     headers: { 'Content-Type': 'application/vnd.api+json; charset=uft8' }
@@ -90,19 +186,28 @@ let arc5 = {
     body: json
   },
 
-  /**
-   * Dependency-free response format
-   */
   // Set type
   type: {
-    body: 'hi there',
+    body: text,
     type: 'text/plain'
   },
 
   // Set cookie
   cookie: {
     body: html,
-    cookie: { hi: 'there' }
+    cookie: 'hi=there'
+  },
+
+  // ... now SSL
+  secureCookie: {
+    body: html,
+    cookie: 'hi=there; Secure'
+  },
+
+  // ... now via header
+  secureCookieHeader: {
+    body: html,
+    headers: { 'set-cookie': 'hi=there; Secure' }
   },
 
   // Set cors
@@ -132,25 +237,25 @@ let arc4 = {
 
 let arc = {
   /**
-   * Arc version agnostic response params
+   * Arc Functions + early Arc response params
    */
   // Set location
-  locationHi: {
+  location: {
     location: '/hi'
   },
 
-  // Set status
-  status201: {
+  // Set statusCode via status
+  status: {
     status: 201
   },
 
-  // Set code
-  code201: {
+  // Set statusCode via code
+  code: {
     code: 201
   },
 
   // Set statusCode
-  statusCode201: {
+  statusCode: {
     statusCode: 201
   },